Refactor Tarjan (#231)
* Place tarjan's graph helpers at graph.py for reuse

* Transforms tarjan algorithm into class

* Add tarjan's examples as test cases

* Comment out testing lines
