+++ graph/test_graph.py
@@ -1,0 +1,44 @@
+from tarjan import Tarjan
+
+import unittest
+
+
+class TestTarjan(unittest.TestCase):
+    """
+    Test for the file tarjan.py
+
+    Arguments:
+        unittest {[type]} -- [description]
+    """
+
+    def test_tarjan_example_1(self):
+        # Graph from https://en.wikipedia.org/wiki/File:Scc.png
+        example = {
+            'A': ['B'],
+            'B': ['C', 'E', 'F'],
+            'C': ['D', 'G'],
+            'D': ['C', 'H'],
+            'E': ['A', 'F'],
+            'F': ['G'],
+            'G': ['F'],
+            'H': ['D', 'G']
+        }
+
+        g = Tarjan(example)
+        self.assertEqual(g.sccs, [['F', 'G'], ['H', 'D', 'C'], ['E', 'B', 'A']])
+
+    def test_tarjan_example_2(self):
+        # Graph from https://en.wikipedia.org/wiki/Tarjan%27s_strongly_connected_components_algorithm#/media/File:Tarjan%27s_Algorithm_Animation.gif
+        example = {
+            'A': ['E'],
+            'B': ['A'],
+            'C': ['B', 'D'],
+            'D': ['C'],
+            'E': ['B'],
+            'F': ['B', 'E', 'G'],
+            'G': ['F', 'C'],
+            'H': ['G', 'H', 'D']
+        }
+
+        g = Tarjan(example)
+        self.assertEqual(g.sccs, [['B', 'E', 'A'], ['D', 'C'], ['G', 'F'], ['H']])
